fetch/metadata/generated/element-video-sec-fetch-user.https.sub.html

<!DOCTYPE html>
<!-- This test was procedurally generated. Please do not modify it directly. -->
<html lang="en">
  <meta charset="utf-8">
  <title>Fetch Metadata - Sec-Fetch-User</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <body>
  <script>
  'use strict';

  function induceRequest(url, attributes) {
    const video = document.createElement('video');

    for (const [ name, value ] of Object.entries(attributes)) {
      video.setAttribute(name, value);
    }

    return new Promise((resolve) => {
        video.setAttribute('src', url);
        video.onload = video.onerror = resolve;
      });
  }

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, []),
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-user');
        });
  }, '');
  </script>
  </body>
</html>
